bot/requirements.txt
python-telegram-bot==21.1.1
django-environ==0.11.2
aiohttp==3.9.3
aiomysql
loguru
cryptography